==> Cyprus_Paper/01_chart_mean_max_etc.ipynb <==
"""
Δημιουργώ τέσσερις ξεχωριστούς πίνακες με τα στοιχεία
mean, min, max, std.
Χρησιμοποιώ το αρχείο o7_patient_chartevent_without_nan
όπου εμπεριέχει τις ωριαίες καταγραφές των ασθενών που έχουν
εισαχθεί στις μονάδες εντατικής θεραπείας
"""

import pandas as pd
time = '3H'

# CSV με τα chartevent των ασθενών
df_patients_chart = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o7_patient_chartevent_without_nan.csv')


# Μετατροπή της στήλης charttime από object σε datetime
df_patients_chart['charttime']= pd.to_datetime(df_patients_chart['charttime'])

# Στοίχιση στοιχείων
df_patients_chart = df_patients_chart.sort_values(by=['hadm_id', 'subject_id', 'charttime'])

df_patients_chart

# Μέση τιμή (mean) ανά την ώρα που έχω ορίσει στην μεταβλητή time 
grouped_multiple = (df_patients_chart.groupby(['subject_id',
                                               'hadm_id',
                                               'itemid',
                                               pd.Grouper(key='charttime',
                                                          freq=time)]).agg({'valuenum': ['mean']}))
grouped_multiple.columns = ['Mean_Chart']
grouped_multiple = grouped_multiple.reset_index()

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
grouped_multiple['date'] = pd.to_datetime(grouped_multiple['charttime']).dt.date
grouped_multiple['time'] = pd.to_datetime(grouped_multiple['charttime']).dt.time

# drop column charttime
grouped_multiple.drop('charttime', axis=1, inplace=True)

# csv export
grouped_multiple.to_csv('D:\Papers\Cyprus_Paper\Data\o8_chartevent_24_hour_mean.csv', index=False)


# Μικρότερη τιμή (min) ανά την ώρα που έχω ορίσει στην μεταβλητή time.
grouped_multiple = (df_patients_chart.groupby(['subject_id',
                                               'hadm_id',
                                               'itemid',
                                               pd.Grouper(key='charttime',
                                                          freq=time)]).agg({'valuenum': ['min']}))
grouped_multiple.columns = ['Min_Chart']
grouped_multiple = grouped_multiple.reset_index()

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
grouped_multiple['date'] = pd.to_datetime(grouped_multiple['charttime']).dt.date
grouped_multiple['time'] = pd.to_datetime(grouped_multiple['charttime']).dt.time

# drop column charttime
grouped_multiple.drop('charttime', axis=1, inplace=True)

# csv export
grouped_multiple.to_csv('D:\Papers\Cyprus_Paper\Data\o9_chartevent_24_hour_min.csv', index=False)


# Μέγιστη τιμή (max) ανά την ώρα που έχω ορίσει στην μεταβλητή time 
grouped_multiple = (df_patients_chart.groupby(['subject_id',
                                               'hadm_id',
                                               'itemid',
                                               pd.Grouper(key='charttime',
                                                          freq=time)]).agg({'valuenum': ['max']}))
grouped_multiple.columns = ['Max_Chart']
grouped_multiple = grouped_multiple.reset_index()

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
grouped_multiple['date'] = pd.to_datetime(grouped_multiple['charttime']).dt.date
grouped_multiple['time'] = pd.to_datetime(grouped_multiple['charttime']).dt.time

# drop column charttime
grouped_multiple.drop('charttime', axis=1, inplace=True)

# csv export
grouped_multiple.to_csv('D:\Papers\Cyprus_Paper\Data\o10_chartevent_24_hour_max.csv', index=False)


# Τυπική απόκλιση (std) ανά την ώρα που έχω ορίσει στην μεταβλητή time
grouped_multiple = (df_patients_chart.groupby(['subject_id',
                                               'hadm_id',
                                               'itemid',
                                               pd.Grouper(key='charttime',
                                                          freq=time)]).agg({'valuenum': ['std']}))
grouped_multiple.columns = ['StdDev_Chart']
grouped_multiple = grouped_multiple.reset_index()

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
grouped_multiple['date'] = pd.to_datetime(grouped_multiple['charttime']).dt.date
grouped_multiple['time'] = pd.to_datetime(grouped_multiple['charttime']).dt.time

# drop column charttime
grouped_multiple.drop('charttime', axis=1, inplace=True)

# csv export
grouped_multiple.to_csv('D:\Papers\Cyprus_Paper\Data\o11_chartevent_24_hour_std.csv', index=False)

df_patients_mean = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o8_chartevent_24_hour_mean.csv')
df_patients_min = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o9_chartevent_24_hour_min.csv')
df_patients_max = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o10_chartevent_24_hour_max.csv')
df_patients_std = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o11_chartevent_24_hour_std.csv')
