==> Cyprus_Paper/02_create_chart_table.ipynb <==
import pandas as pd

# Καταμέτρηση των εξετάσεων

without_nan = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o7_patient_chartevent_without_nan.csv')

# Καταμέτρηση των εξετάσεων που κάνουν οι ασθενείς με εγκεφαλικό και μετατροπή από series σε dataframe.
rslt_df = pd.DataFrame(without_nan['itemid'].value_counts(dropna=True))

# Το παραγόμενο αρχείο έχει σαν index την στήλη που περιέχει τον κωδικό των εξετάσεων, οπότε την κάνουμε reset
rslt_df = rslt_df.reset_index()

# Ορίζουμε νέες ονομασίες στις στήλες για να μπορούν να είναι ευανάγνωστες.
rslt_df = rslt_df.rename(columns={'index': 'itemid', 'itemid': 'times'})

# Εξαγωγή σε αρχείο
rslt_df.to_csv('D:\Papers\Cyprus_Paper\Data\o12_chartevents_item_id_counts.csv', index=False)

# Δημιουργία του κενού πίνακα mean chartevents

"""
Δημιουργία του πίνακα chartevet mean
με το είδος των εξετάσων διαχωρισμένες
σε 707 στήλες.
"""

#Δημιουργώ τις στήλες από το αρχείο που περιλαμβάνει τις μοναδικές εξετάσεις
df_columns = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o12_chartevents_item_id_counts.csv',
                         usecols = ['itemid'], index_col=False)

#Ορίζω το header ως index έτσι ώστε όταν αλλάξω την κατεύθυνση του csv να μπουν τα στοιχεία στην κεφαλίδα.
df_columns = df_columns.set_index('itemid')

# Μεταφέρω τις γραμμές σε στήλες
df_columns = df_columns.transpose()

df_rows = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o8_chartevent_24_hour_mean.csv')

"""
Δημιουργώ ένα μοναδικό αναγνωριστικό με την ένωση των
στοιχείων από στήλες subject, hadm_id, και charttime.
Ανάμεσα στα στοιχεία προβάλω ένα κενό το οποίο θα μου
φανεί χρήσιμο αργότερα για τον διαχωρισμό τους ξανά σε
διαφορετικές στήλες.
"""
df_rows["concat"] = (df_rows["subject_id"].astype(str)
                     +" "
                     +df_rows["hadm_id"].astype(str)
                     +" "
                     +df_rows["date"].astype(str)
                     +" "
                     +df_rows["time"].astype(str))

# Ελέγχω αν υπάρχει κενό κελί στην στήλη concat.
df_rows['concat'].isnull().values.any()

# Αλλάζω την σειρά των στηλών του dataframe
df_rows = df_rows[['concat', 'subject_id', 'hadm_id', 'date', 'time', 'itemid', 'Mean_Chart']]

# Από το dr_rows κρατάω τις μοναδικές εγγραφές
df_copy_rows = pd.DataFrame(df_rows.concat.unique())

# Δημιουργώ καινούργιο πίνακα ενσωματώνοντας στον πίνακα df columns τις στήλες με τις κενές για την ώρα εξετάσεις
chart_tbl = pd.concat([df_copy_rows, df_columns], axis=1)

# Μετονομάζω την στήλη 0 σε concat για να είναι ίδιες στους δυο πίνακες
chart_tbl = chart_tbl.rename({0:'concat'}, axis=1)

#Διαχωρισμός του concat σε τέσσερις στήλες
chart_tbl[['subject_id','hadm_id', 'date', 'time']] = chart_tbl['concat'].str.split(expand=True)

"""
Διαγράφω την στήλη concat γιατί θα
πρέπει να την ξαναδημιουργήσω χωρίς
κενά ανάμεσα στα ενοποιημένα στοιχεία,
οπότε θα την έχω διπλή.
"""
chart_tbl = chart_tbl.drop('concat', 1)

#Φέρνω την τελευταία στήλη, πρώτη
for i in range (1, 5):
    cols = list(chart_tbl.columns)
    cols = [cols[-1]] + cols[:-1]
    chart_tbl = chart_tbl[cols]
    
chart_tbl.to_csv('D:\Papers\Cyprus_Paper\Data\o13_empty_chart_mean_tbl.csv', index=False)

# Δημιουργία του κενού πίνακα min chartevents

#Δημιουργώ τις στήλες από το αρχείο που περιλαμβάνει τις μοναδικές εξετάσεις
df_columns = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o12_chartevents_item_id_counts.csv',
                         usecols = ['itemid'], index_col=False)

#Ορίζω το header ως index έτσι ώστε όταν αλλάξω την κατεύθυνση του csv να μπουν τα στοιχεία στην κεφαλίδα.
df_columns = df_columns.set_index('itemid')

# Μεταφέρω τις γραμμές σε στήλες
df_columns = df_columns.transpose()

df_rows = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o9_chartevent_24_hour_min.csv')

"""
Δημιουργώ ένα μοναδικό αναγνωριστικό με την ένωση των
στοιχείων από στήλες subject, hadm_id και charttime.
Ανάμεσα στα στοιχεία προβάλω ένα κενό το οποίο θα μου
φανεί χρήσιμο αργότερα για τον διαχωρισμό τους ξανά σε
διαφορετικές στήλες.
"""
df_rows["concat"] = (df_rows["subject_id"].astype(str)
                     +" "
                     +df_rows["hadm_id"].astype(str)
                     +" "
                     +df_rows["date"].astype(str)
                     +" "
                     +df_rows["time"].astype(str))

# Ελέγχω αν υπάρχει κενό κελί στην στήλη concat.
df_rows['concat'].isnull().values.any()

# Αλλάζω την σειρά των στηλών του dataframe
df_rows = df_rows[['concat', 'subject_id', 'hadm_id', 'date', 'time', 'itemid', 'Min_Chart']]

# Από το dr_rows κρατάω τις μοναδικές εγγραφές
df_copy_rows = pd.DataFrame(df_rows.concat.unique())

# Δημιουργώ καινούργιο πίνακα ενσωματώνοντας στον πίνακα df columns τις στήλες με τις κενές για την ώρα εξετάσεις
chart_tbl = pd.concat([df_copy_rows, df_columns], axis=1)

# Μετονομάζω την στήλη 0 σε concat για να είναι ίδιες στους δυο πίνακες
chart_tbl = chart_tbl.rename({0:'concat'}, axis=1)

#Διαχωρισμός του concat σε τέσσερις στήλες
chart_tbl[['subject_id','hadm_id', 'date', 'time']] = chart_tbl['concat'].str.split(expand=True)

"""
Διαγράφω την στήλη concat γιατί θα
πρέπει να την ξαναδημιουργήσω χωρίς
κενά ανάμεσα στα ενοποιημένα στοιχεία,
οπότε θα την έχω διπλή.
"""
chart_tbl = chart_tbl.drop('concat', 1)

#Φέρνω την τελευταία στήλη, πρώτη
for i in range (1, 5):
    cols = list(chart_tbl.columns)
    cols = [cols[-1]] + cols[:-1]
    chart_tbl = chart_tbl[cols]
    
chart_tbl.to_csv('D:\Papers\Cyprus_Paper\Data\o14_empty_chart_min_tbl.csv', index=False)

# Δημιουργία του κενού πίνακα max chartevents

#Δημιουργώ τις στήλες από το αρχείο που περιλαμβάνει τις μοναδικές εξετάσεις
df_columns = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o12_chartevents_item_id_counts.csv',
                         usecols = ['itemid'], index_col=False)

#Ορίζω το header ως index έτσι ώστε όταν αλλάξω την κατεύθυνση του csv να μπουν τα στοιχεία στην κεφαλίδα.
df_columns = df_columns.set_index('itemid')

# Μεταφέρω τις γραμμές σε στήλες
df_columns = df_columns.transpose()

df_rows = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o10_chartevent_24_hour_max.csv')

"""
Δημιουργώ ένα μοναδικό αναγνωριστικό με την ένωση των
στοιχείων από στήλες subject, hadm_id και charttime.
Ανάμεσα στα στοιχεία προβάλω ένα κενό το οποίο θα μου
φανεί χρήσιμο αργότερα για τον διαχωρισμό τους ξανά σε
διαφορετικές στήλες.
"""
df_rows["concat"] = (df_rows["subject_id"].astype(str)
                     +" "
                     +df_rows["hadm_id"].astype(str)
                     +" "
                     +df_rows["date"].astype(str)
                     +" "
                     +df_rows["time"].astype(str))

# Ελέγχω αν υπάρχει κενό κελί στην στήλη concat.
df_rows['concat'].isnull().values.any()

# Αλλάζω την σειρά των στηλών του dataframe
df_rows = df_rows[['concat', 'subject_id', 'hadm_id', 'date', 'time', 'itemid', 'Max_Chart']]

# Από το dr_rows κρατάω τις μοναδικές εγγραφές
df_copy_rows = pd.DataFrame(df_rows.concat.unique())

# Δημιουργώ καινούργιο πίνακα ενσωματώνοντας στον πίνακα df columns τις στήλες με τις κενές για την ώρα εξετάσεις
chart_tbl = pd.concat([df_copy_rows, df_columns], axis=1)

# Μετονομάζω την στήλη 0 σε concat για να είναι ίδιες στους δυο πίνακες
chart_tbl = chart_tbl.rename({0:'concat'}, axis=1)

#Διαχωρισμός του concat σε τέσσερις στήλες
chart_tbl[['subject_id','hadm_id', 'date', 'time']] = chart_tbl['concat'].str.split(expand=True)

"""
Διαγράφω την στήλη concat γιατί θα
πρέπει να την ξαναδημιουργήσω χωρίς
κενά ανάμεσα στα ενοποιημένα στοιχεία,
οπότε θα την έχω διπλή.
"""
chart_tbl = chart_tbl.drop('concat', 1)

#Φέρνω την τελευταία στήλη, πρώτη
for i in range (1, 5):
    cols = list(chart_tbl.columns)
    cols = [cols[-1]] + cols[:-1]
    chart_tbl = chart_tbl[cols]
    
chart_tbl.to_csv('D:\Papers\Cyprus_Paper\Data\o15_empty_chart_max_tbl.csv', index=False)

# Δημιουργία του κενού πίνακα Standard Deviation chartevents

#Δημιουργώ τις στήλες από το αρχείο που περιλαμβάνει τις μοναδικές εξετάσεις
df_columns = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o12_chartevents_item_id_counts.csv',
                         usecols = ['itemid'], index_col=False)

#Ορίζω το header ως index έτσι ώστε όταν αλλάξω την κατεύθυνση του csv να μπουν τα στοιχεία στην κεφαλίδα.
df_columns = df_columns.set_index('itemid')

# Μεταφέρω τις γραμμές σε στήλες
df_columns = df_columns.transpose()

df_rows = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o11_chartevent_24_hour_std.csv')

"""
Δημιουργώ ένα μοναδικό αναγνωριστικό με την ένωση των
στοιχείων από στήλες subject, hadm_id και charttime.
Ανάμεσα στα στοιχεία προβάλω ένα κενό το οποίο θα μου
φανεί χρήσιμο αργότερα για τον διαχωρισμό τους ξανά σε
διαφορετικές στήλες.
"""
df_rows["concat"] = (df_rows["subject_id"].astype(str)
                     +" "
                     +df_rows["hadm_id"].astype(str)
                     +" "
                     +df_rows["date"].astype(str)
                     +" "
                     +df_rows["time"].astype(str))

# Ελέγχω αν υπάρχει κενό κελί στην στήλη concat.
df_rows['concat'].isnull().values.any()

# Αλλάζω την σειρά των στηλών του dataframe
df_rows = df_rows[['concat', 'subject_id', 'hadm_id', 'date', 'time', 'itemid', 'StdDev_Chart']]

# Από το dr_rows κρατάω τις μοναδικές εγγραφές
df_copy_rows = pd.DataFrame(df_rows.concat.unique())

# Δημιουργώ καινούργιο πίνακα ενσωματώνοντας στον πίνακα df columns τις στήλες με τις κενές για την ώρα εξετάσεις
chart_tbl = pd.concat([df_copy_rows, df_columns], axis=1)

# Μετονομάζω την στήλη 0 σε concat για να είναι ίδιες στους δυο πίνακες
chart_tbl = chart_tbl.rename({0:'concat'}, axis=1)

#Διαχωρισμός του concat σε τέσσερις στήλες
chart_tbl[['subject_id','hadm_id', 'date', 'time']] = chart_tbl['concat'].str.split(expand=True)

"""
Διαγράφω την στήλη concat γιατί θα
πρέπει να την ξαναδημιουργήσω χωρίς
κενά ανάμεσα στα ενοποιημένα στοιχεία,
οπότε θα την έχω διπλή.
"""
chart_tbl = chart_tbl.drop('concat', 1)

#Φέρνω την τελευταία στήλη, πρώτη
for i in range (1, 5):
    cols = list(chart_tbl.columns)
    cols = [cols[-1]] + cols[:-1]
    chart_tbl = chart_tbl[cols]
    
chart_tbl.to_csv('D:\Papers\Cyprus_Paper\Data\o16_empty_chart_std_tbl.csv', index=False)
