==> Cyprus_Paper/05_fill_max_chart_table.ipynb <==
import pandas as pd

df_chart_tbl = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o15_empty_chart_max_tbl.csv', index_col=False)
df_max_chart = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o10_chartevent_24_hour_max.csv', index_col=False)
df_data_chart = pd.read_csv('D:\MSc Aegean\Thesis\mimic-iv-1.0\icu\d_items.csv.gz')

"""
Δημιουργώ εκ νέου την στήλη concat
χωρίς κενά ανάμεσα στα στοιχεία.
Θα χρησιμοποιήσω αυτή τη στήλη σαν
index γιατί τα κενά μπορεί να
δημιουργήσουν πρόβλημα.
"""
df_chart_tbl["concat"] = (df_chart_tbl["subject_id"].astype(str)
                          + df_chart_tbl["hadm_id"].astype(str)
                          + df_chart_tbl["date"].astype(str)
                          + df_chart_tbl["time"].astype(str))

"""
Αφαιρώ τους χαρακτήρες - και :
από της τιμές της concat για να
αποφύγω πιθανά προβλήματα
"""
df_chart_tbl['concat'] = df_chart_tbl['concat'].str.replace('-','')
df_chart_tbl['concat'] = df_chart_tbl['concat'].str.replace(':','')

# Ορίζω την στήλη concat ως index των γραμμών
df_chart_tbl = df_chart_tbl.set_index('concat')

"""
Δημιουργώ την στήλη concat και στον πίνακα
που περιλαμβάνει τις μετρήσεις για να είναι
το κοινό σημείο ανάμεσα στους δυο πίνακες
"""
df_max_chart["concat"] = (df_max_chart["subject_id"].astype(str)
                          + df_max_chart["hadm_id"].astype(str)
                          + df_max_chart["date"].astype(str)
                          + df_max_chart["time"].astype(str))

"""
Αφαιρώ τους χαρακτήρες - και :
από της τιμές της concat για να
αποφύγω πιθανά προβλήματα
"""
df_max_chart['concat'] = df_max_chart['concat'].str.replace('-','')
df_max_chart['concat'] = df_max_chart['concat'].str.replace(':','')

df_max_chart

# Γεμίζω τον πίνακα
for row in df_max_chart.itertuples():
    con = row.concat
    item = row.itemid
    value = row.Max_Chart
    df_chart_tbl.at['{}'.format(con), '{}'.format(item)] = value

# Αποθήκευση του τελικού αρχείου με επικεφαλίδες τους κωδικούς των παρατηρήσεων
df_chart_tbl.to_csv('D:\Papers\Cyprus_Paper\Data\o21_fill_chart_max_itemid_header_tbl.csv', index=False)

df_data_chart = df_data_chart.set_index('itemid')

# Ενοποίηση τιμών σε κελιά
df_data_chart["header"] = ("Max - "
                           + df_data_chart["label"].astype(str)
                           + " ("
                           + df_data_chart["unitname"].astype(str)
                           + ")")

"""
Ελέγχω τις κεφαλίδες και αν δεν ανήκουν στις
τέσσερις πρώτες τιμές κάνω τις αντικατάσταση.
"""
for col in df_chart_tbl.columns:
    if ((col != 'subject_id') and
        (col != 'hadm_id') and
        (col != 'date') and
        (col != 'time')):
        temp = (df_data_chart.loc[int(col), 'header'])
        df_chart_tbl.rename(columns={'{}'.format(col): '{}'.format(temp)}, inplace=True)

df_chart_tbl

# Αποθήκευση του τελικού αρχείου με επικεφαλίδες τις ονομασίες των παρατηρήσεων
df_chart_tbl.to_csv('D:\Papers\Cyprus_Paper\Data\o22_fill_max_chart_label_header_tbl.csv', index=False)
