==> Cyprus_Paper/08_lab_mean_max_etc.ipynb <==
"""
Δημιουργώ τέσσερις ξεχωριστούς πίνακες με τα στοιχεία
mean, min, max, std.
Χρησιμοποιώ το αρχείο o7_patient_chartevent_without_nan
όπου εμπεριέχει τις ωριαίες καταγραφές των ασθενών που έχουν
εισαχθεί στις μονάδες εντατικής θεραπείας
"""

import pandas as pd
time = '96H'

# CSV με τα chartevent των ασθενών
df_labevents= pd.read_csv('D:\MSc Aegean\Thesis\exports\o13_icu_labevents.csv')

df_labevents

# Μετατροπή της στήλης charttime από object σε datetime

df_labevents['charttime']= pd.to_datetime(df_labevents['charttime'])

#Απομονώνω τις γραμμές που δεν περιλαμβάνουν τιμές στην valuenum
df_labevents = df_labevents[df_labevents['valuenum'].notna()]

# Μέση τιμή (mean) ανά 24ώρο 
grouped_multiple = (df_labevents.groupby(['subject_id',
                                          'hadm_id',
                                          'itemid',
                                          pd.Grouper(key='charttime',
                                                     freq=time)]).agg({'valuenum': ['mean']}))
grouped_multiple.columns = ['Mean_Lab']
grouped_multiple = grouped_multiple.reset_index()

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
grouped_multiple['date'] = pd.to_datetime(grouped_multiple['charttime']).dt.date
grouped_multiple['time'] = pd.to_datetime(grouped_multiple['charttime']).dt.time

# drop column charttime
grouped_multiple.drop('charttime', axis=1, inplace=True)

grouped_multiple.to_csv('D:\Papers\Cyprus_Paper\Data\o26_labevents_24hours_mean.csv', index=False)



# Μικρότερη τιμή (min) ανά 24ώρο 
grouped_multiple = (df_labevents.groupby(['subject_id',
                                          'hadm_id',
                                          'itemid',
                                          pd.Grouper(key='charttime',
                                                     freq=time)]).agg({'valuenum': ['min']}))
grouped_multiple.columns = ['Min_Lab']
grouped_multiple = grouped_multiple.reset_index()

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
grouped_multiple['date'] = pd.to_datetime(grouped_multiple['charttime']).dt.date
grouped_multiple['time'] = pd.to_datetime(grouped_multiple['charttime']).dt.time

# drop column charttime
grouped_multiple.drop('charttime', axis=1, inplace=True)

grouped_multiple.to_csv('D:\Papers\Cyprus_Paper\Data\o27_labevents_24_hour_min.csv', index=False)


# Μέγιστη τιμή (max) ανά 24ώρο 
grouped_multiple = (df_labevents.groupby(['subject_id',
                                          'hadm_id',
                                          'itemid',
                                          pd.Grouper(key='charttime',
                                                     freq=time)]).agg({'valuenum': ['max']}))
grouped_multiple.columns = ['Max_Lab']
grouped_multiple = grouped_multiple.reset_index()

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
grouped_multiple['date'] = pd.to_datetime(grouped_multiple['charttime']).dt.date
grouped_multiple['time'] = pd.to_datetime(grouped_multiple['charttime']).dt.time

# drop column charttime
grouped_multiple.drop('charttime', axis=1, inplace=True)

grouped_multiple.to_csv('D:\Papers\Cyprus_Paper\Data\o28_labevents_24_hour_max.csv', index=False)


# Τυπική απόκλιση (std) ανά 24ώρο 
grouped_multiple = (df_labevents.groupby(['subject_id',
                                          'hadm_id',
                                          'itemid',
                                          pd.Grouper(key='charttime',
                                                     freq=time)]).agg({'valuenum': ['std']}))
grouped_multiple.columns = ['StdDev_Lab']
grouped_multiple = grouped_multiple.reset_index()

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
grouped_multiple['date'] = pd.to_datetime(grouped_multiple['charttime']).dt.date
grouped_multiple['time'] = pd.to_datetime(grouped_multiple['charttime']).dt.time

# drop column charttime
grouped_multiple.drop('charttime', axis=1, inplace=True)

grouped_multiple.to_csv('D:\Papers\Cyprus_Paper\Data\o29_labevent_24_hour_std.csv', index=False)

df_patients_mean = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o26_labevents_24hours_mean.csv')
df_patients_min = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o27_labevents_24_hour_min.csv')
df_patients_max = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o28_labevents_24_hour_max.csv')
df_patients_std = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o29_labevent_24_hour_std.csv')
