==> Cyprus_Paper/14_merge_lab_tables.ipynb <==
import pandas as pd

df_mean = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o36_fill_mean_lab_label_header_tbl.csv')

df_min = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o38_fill_min_lab_label_header_tbl.csv')

df_max = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o41_fill_max_lab_label_header_tbl.csv')

df_std = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o43_fill_std_lab_label_header_tbl.csv')

# Συγχώνευση πρώτων δύο πινάκων

#Ελέγχω αν μετά την συγχώνευση υπάρχουν διπλές ίδιες στήλες

rslt=pd.merge(df_mean, df_min, on=['subject_id','hadm_id', 'date', 'time'],
              how="outer", suffixes=('', '_delme'))

# Και αν υπάρχουν τις αφαιρώ
rslt = rslt[[c for c in rslt.columns if not c.endswith('_delme')]]

# Συγχώνευση των δύο πρώτων πινάκων με τον τρίτο

#Ελέγχω αν μετά την συγχώνευση υπάρχουν διπλές ίδιες στήλες

rslt=pd.merge(rslt, df_max, on=['subject_id','hadm_id', 'date', 'time'],
              how="outer", suffixes=('', '_delme'))

# Και αν υπάρχουν τις αφαιρώ
rslt = rslt[[c for c in rslt.columns if not c.endswith('_delme')]]

# Συγχώνευση των τριών πρώτων πινάκων με τον τέταρτο

#Ελέγχω αν μετά την συγχώνευση υπάρχουν διπλές ίδιες στήλες

rslt=pd.merge(rslt, df_std, on=['subject_id','hadm_id', 'date', 'time'],
              how="outer", suffixes=('', '_delme'))

# Και αν υπάρχουν τις αφαιρώ
rslt = rslt[[c for c in rslt.columns if not c.endswith('_delme')]]

rslt.sort_values(by=['subject_id', 'date', 'time'])

#Εξαγωγή σε csv
rslt.to_csv('D:\Papers\Cyprus_Paper\Data\o44_merged_lab_tables.csv', index=False)
