==> Cyprus_Paper/15_Merge_Chart_Lab.ipynb <==
import pandas as pd

df_chart = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o25_merged_chart_tables.csv')

df_lab = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o44_merged_lab_tables.csv')

#Ελέγχω αν μετά την συγχώνευση υπάρχουν διπλές ίδιες στήλες

rslt=pd.merge(df_chart, df_lab, on=['subject_id','hadm_id', 'date', 'time'],
              how="outer", suffixes=('', '_delme'))

# Και αν υπάρχουν τις αφαιρώ
rslt = rslt[[c for c in rslt.columns if not c.endswith('_delme')]]

rslt = rslt.sort_values(by=['subject_id', 'date', 'time'])

#Εξαγωγή σε csv
rslt.to_csv('D:\Papers\Cyprus_Paper\Data\o45_merged_chart_lab.csv', index=False)
