==> Cyprus_Paper/16_set_death_flag.ipynb <==
import pandas as pd
import numpy as np

# Εισαγωγή Expire flag

# Εισαγωγή csv
df_icu = pd.read_csv('D:\Papers\Cyprus_Paper\Data\o45_merged_chart_lab.csv')

df_patients = pd.read_csv('D:\MSc Aegean\Thesis\mimic-iv-1.0\core\patients.csv.gz')

df_admissions = pd.read_csv('D:\MSc Aegean\Thesis\mimic-iv-1.0\core\Admissions.csv.gz', usecols = ['subject_id', 'hadm_id', 'hospital_expire_flag'])

df_death = df_admissions.loc[df_admissions['hospital_expire_flag'] == 1]

#Ελέγχω αν μετά την συγχώνευση υπάρχουν διπλές ίδιες στήλες
df_rslt=pd.merge(df_icu, df_death, on=['subject_id','hadm_id'],
              how="left", suffixes=('', '_delme'))


# Και αν υπάρχουν τις αφαιρώ
df_rslt = df_rslt[[c for c in df_rslt.columns if not c.endswith('_delme')]]

# Εισάγω την τιμή μηδέν στα κενά κελιά της hospital_expire_flag
df_rslt[['hospital_expire_flag']] = df_rslt[['hospital_expire_flag']].fillna(0)

# Μετατρέπω την στήλη hospital_expire_flag σε ακέραιο.
df_rslt['hospital_expire_flag'] = df_rslt['hospital_expire_flag'].apply(int)

"""
Αντικατάσταση τιμών από μηδέν σε
survival και από ένα σε death
στην στήλη hospital_expire_flag
"""

df_rslt["hospital_expire_flag"] = np.where(df_rslt["hospital_expire_flag"] == 0, "survival", "death")

#Εξαγωγή σε csv
df_rslt.to_csv('D:\Papers\Cyprus_Paper\Data\o46_chart_lab_with_death_flag.csv', index=False)
