==> Cyprus_Paper/20_Setting_Survival_Days.ipynb <==
import pandas as pd
import numpy as np

# Διαβάζω τα csv
df_icu_stay = pd.read_csv('D:\MSc Aegean\Thesis\mimic-iv-1.0\icu\icustays.csv.gz', usecols = ['subject_id', 'hadm_id', 'los'])
df_hours = pd.read_csv('D:\Papers\Cyprus_Paper\Data\Twenty_four_hours\o3 hours\o1_3hours.csv')
df_hours = df_hours.drop(['Survival Days'], axis=1)

'''
Ομαδοποίηση ημερών παραμονής στη ΜΕΘ
μιας και υπάρχουν πολλαπλές εγγραφές
λόγο μετακινήσεων με το ίδιο hadm_id.
Έτσι παίρνω το άθροισμα των ημερών
παραμονής.
'''
df_temp = df_icu_stay.groupby(df_icu_stay['hadm_id']).aggregate({'subject_id': 'first', 'hadm_id': 'first', 'los': 'sum'}).reset_index(drop=True)

# Ελέγχω αν μετά την συγχώνευση υπάρχουν διπλές ίδιες στήλες

rslt_df=pd.merge(df_hours, df_temp, on=['subject_id','hadm_id'],
              how="left", suffixes=('', '_delme'))

# Και αν υπάρχουν τις αφαιρώ
rslt_df = rslt_df[[c for c in rslt_df.columns if not c.endswith('_delme')]]

# Αντικαθιστώ τις τιμές στην στήλη Los στους επιζήσαντες με το 365
rslt_df.loc[rslt_df.hospital_expire_flag == 'survival', 'los'] = 365

# Μετενομ΄άζω την στήλη los σε Survival Days
rslt_df = rslt_df.rename(columns={"los": "Survival Days"})

rslt_df

# αντικατάσταση του survival και death με 0 & 1
rslt_df["hospital_expire_flag"].replace({"survival": 0, "death": 1}, inplace=True)

rslt_df.to_csv('D:\Papers\Cyprus_Paper\Data\Twenty_four_hours\o3 hours\o1_3hours.csv', index=False)
